Split toolchain test helpers by responsibility (#207)
Derive the expected component list in the component-install failure test
from `REQUIRED_COMPONENTS` so the assertion stays aligned with the
production constant.

Move the command-capture helper into `test_helpers.rs`, extract the
failure-specific setup and assertion helpers into `failure_mocks.rs`,
and remove the partially-adopted channel newtype so the test module uses
plain `&str` channels consistently.
